+++ assets/js/components/VehicleIcon/Cooking.vue
@@ -7,8 +7,10 @@
 	</svg>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from "vue";
+
+export default defineComponent({
 	name: "Cooking",
-};
+});
 </script>
