+++ assets/js/components/VehicleIcon/Heatexchange.vue
@@ -7,8 +7,9 @@
 	</svg>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from "vue";
+export default defineComponent({
 	name: "Heatexchange",
-};
+});
 </script>
