+++ assets/js/components/VehicleIcon/Kettle.vue
@@ -7,8 +7,9 @@
 	</svg>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from "vue";
+export default defineComponent({
 	name: "Kettle",
-};
+});
 </script>
