+++ assets/js/components/VehicleIcon/Rocket.vue
@@ -7,8 +7,10 @@
 	</svg>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from "vue";
+
+export default defineComponent({
 	name: "Rocket",
-};
+});
 </script>
