+++ assets/js/components/VehicleIcon/Shuttle.vue
@@ -7,8 +7,10 @@
 	</svg>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from "vue";
+
+export default defineComponent({
 	name: "Shuttle",
-};
+});
 </script>
