+++ assets/js/components/VehicleIcon/index.ts
